pyproject.toml
[project]
name = "oc-p10-softdesk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.4.0",
]
